+++ tests/compile-and-dump/Singletons/Error.hs
@@ -1,7 +1,7 @@
 module Singletons.Error where
 
 import Data.Singletons
-import Data.Singletons.Prelude hiding (Head, HeadSym0, HeadSym1)
+import Data.Singletons.Prelude hiding (Head, HeadSym0, HeadSym1, sHead)
 import Data.Singletons.TH
 
 $(singletons [d|